Enable sending the existing settings to caller.
A user of the GPSBeaconThread can now request the existing
gps settings for beacon type, polling, etc.
This is used by the webchat server to request the running
settings so when a user reloads the page, the ui can reflect
the running settings, which may be different from the config
settings.  We allow webchat users to alter the settings on the
fly.